Updated LICENSE, Argument parsing, and tests
Updated the license to include the text of the licenses from used
libraries. Added references to libraries to the README, and added a
short description. Fixed bugs with _mi_cont_cont_gen, and the tests for
this function. Added tests for _mi_cont_cont_cheb_only (not passing).
Created functions for easier matching of string arguments against groups
of arguments.

TODO: Fix tests for Chebyshev only MI function as well as remaining
mutual information functions.