Added Methods for Calculating the Jensen-Shannon Divergence
Added methods for calculating the Jensen-Shannon Divergence between two
discrete or two continuous distributions. Added a _main_wrapper helper
in the divergence submodule to help with wrapping the discrete and
continuous cases in a consistent way between Jensen-Shannon and
Kullback-Leibler divergence methods. Refactored the Kullback-Leibler
Divergence methods to make use of this new wrapper. Modifications to
comments in mutual_information_functions.

TODO: Add tests for the kl_divergence_functions, and the js_divergence_functions