Updated Model Creation Methods
Updated the model creation methods in the model_creation module. The fva
method now only runs fva for the reactions which are over or underexpressed,
and the milp method now handles errors slightly differently. For the
milp method, it now will select the best of forward, reverse, or
inactive, treating any infeasible solutions as worse than other
solutions, rather than not picking a case when one of the cases is
infeasible.