Added method for Density Calculation and Bipartite Projection
Added methods for calculating density of node labels on a graph,
as well as tests for this. Added methods for converting between gene
and reaction identifiers of various forms (mapping using cobra
model), as well as tests for them. Added methods to project
a bipartite graph, but no tests added for these methods yet.