Updated Docs Added Metabolite Methods
Added methods for metabolite focused FBA. Specifically
added functions for finding metabolite networks, and
for running the metchange algorithms (as well as some
context managers to help with these functions).

Added tests for the metabolite network functions,
need to add tests for the metchange functions.

Updated the documentation to include these metabolite
focused methods, and renamed graph to network.