Added CLI script for running Metchange
Added CLI script for running the metchange algorithm, and tests for this
script. Updated names of expression utils functions to reflect which
algorithm they should be applied with. Added function dictionary globals
to the GPR module, so that they don't have to be copied by hand to use.
Should create a configuration module/object in future so that there is
less independendent declarations of default values (or at least, all the
different defaults are in a single file).