Updated imatgen script to include Mutli and Single Modes
Updated the imatgen script so that in addition to being able to create
single IMAT models, it can create multiple IMAT models sequentially.
Also included the ability to use fold change to define gene expression
weights rather than the default quantiles. Updated imatgen tests (for
the single case) to reflect changed usage.

Updated numpy usage to be consistent with 2.0 release, but cobrapy is
not yet updated to work with numpy 2.0 release, so that update in
dependencies will need to wait.

TODO: Add tests for imatgen multi mode