Added pre-commit-config and Fixed Translate Tests
Set up pre-commit to run hooks for trailing-whitespace,
end-of-file-fixer, check-yaml, check-added-large-files, ruff lint, and
ruff format.

Fixed tests for utils.translate so that the order of the returned
gene names is not important (since the use of sets to ensure uniqueness
of returned genes means that order is not garaunteed).