Release of version 0.2.1
Releasing version 0.2.1 with some added functionality, primarily the
grouped divergence and the pairwise mutual information. Grouped
divergence enables calculating the divergence between two dataframes (or
numpy arrays) for different groups of columns, while pairwise mutual
information allows for calculating the mutual information between every
pair of columns in a dataframe (or numpy array). None of the changes
should be breaking, any that are are considered bugs, and will be fixed.